templates migration to ts
