+++ src/utils/time.ts
@@ -1,8 +1,9 @@
 import { DateTime } from "luxon"
+import { Lang } from "./languages"
 
 export const INVALID_DATETIME = "Invalid DateTime"
 
-export const getLocaleTimestamp = (locale, timestamp) => {
+export const getLocaleTimestamp = (locale: Lang, timestamp: string) => {
   let localeTimestamp = DateTime.fromSQL(timestamp)
     .setLocale(locale)
     .toLocaleString(DateTime.DATE_FULL)
